requirements-tests.txt
django-stubs
mypy == 1.13.0
pytest
pytest-asyncio
pytest-cov
pytest-django
ruff == 0.8.4
